Fix review issues in Pest migration
