+++ composer.json
@@ -27,8 +27,8 @@
         "laravel/pint": "^1.26",
         "mockery/mockery": "^1.6.12",
         "orchestra/testbench": "^10.6|^11.0",
-        "pestphp/pest": "*",
-        "pestphp/pest-plugin-laravel": "*"
+        "pestphp/pest": "^3.0|^4.0",
+        "pestphp/pest-plugin-laravel": "^3.0|^4.0"
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
