+++ tests/Feature/Providers/Gemini/BaseUrlTest.php
@@ -2,9 +2,6 @@
 
 use Illuminate\Support\Facades\Http;
 use Tests\Feature\Agents\AssistantAgent;
-use Tests\Feature\Providers\Gemini\GeminiHelpers;
-
-uses(GeminiHelpers::class);
 
 test('gemini requests use the configured base url', function () {
     config(['ai.providers.gemini.url' => 'https://custom-proxy.example.com/v1']);
