+++ tests/Feature/Providers/Gemini/ToolCallLoopTest.php
@@ -3,9 +3,6 @@
 use GuzzleHttp\Promise\PromiseInterface;
 use Illuminate\Support\Facades\Http;
 use Tests\Feature\Agents\ToolUsingAgent;
-use Tests\Feature\Providers\Gemini\GeminiHelpers;
-
-uses(GeminiHelpers::class);
 
 test('tool calls trigger follow up request', function () {
     Http::fake([
