+++ tests/Feature/Providers/Groq/RequestMappingTest.php
@@ -1,6 +1,5 @@
 <?php
 
-use GuzzleHttp\Promise\PromiseInterface;
 use Illuminate\Http\Client\Request;
 use Illuminate\Support\Facades\Http;
 use Tests\Feature\Agents\AssistantAgent;
@@ -192,24 +191,3 @@ test('structured response is correctly parsed', function () {
 
     expect($response->structured['symbol'])->toBe('Au');
 });
-
-function fakeGroqResponse(string $text): PromiseInterface
-{
-    return Http::response([
-        'id' => 'chatcmpl-123',
-        'object' => 'chat.completion',
-        'model' => 'openai/gpt-oss-20b',
-        'choices' => [[
-            'index' => 0,
-            'message' => [
-                'role' => 'assistant',
-                'content' => $text,
-            ],
-            'finish_reason' => 'stop',
-        ]],
-        'usage' => [
-            'prompt_tokens' => 1,
-            'completion_tokens' => 1,
-        ],
-    ]);
-}
