+++ tests/Feature/Providers/OpenAi/ToolMappingTest.php
@@ -1,6 +1,5 @@
 <?php
 
-use GuzzleHttp\Promise\PromiseInterface;
 use Illuminate\Http\Client\Request;
 use Illuminate\Support\Facades\Http;
 use Tests\Feature\Tools\FixedNumberGenerator;
@@ -55,24 +54,3 @@ test('tool with empty schema includes strict compliant parameters', function () 
             && $tool['parameters']['additionalProperties'] === false;
     });
 });
-
-function fakeOpenAiResponse(string $text): PromiseInterface
-{
-    return Http::response([
-        'id' => 'resp_123',
-        'status' => 'completed',
-        'model' => 'gpt-5.4',
-        'output' => [[
-            'type' => 'message',
-            'status' => 'completed',
-            'content' => [[
-                'type' => 'output_text',
-                'text' => $text,
-            ]],
-        ]],
-        'usage' => [
-            'input_tokens' => 1,
-            'output_tokens' => 1,
-        ],
-    ]);
-}
